Refactored QR generator code into separate class
