+++ qr-scan-ios/Controller/GenerateQRViewController.swift
@@ -21,23 +21,9 @@ class GenerateQRViewController: UIViewController {
         NetworkController().loadQRCodeSeed { (seed) in
             guard let seedData = seed else { return }
             DispatchQueue.main.async {
-                self.qrCodeImageView.image = self.generateQRCode(from: seedData)
+                self.qrCodeImageView.image = QRGenerator().generateQRCode(from: seedData)
                 self.spinner.stopAnimating()
             }
         }
-    }
-    
-    func generateQRCode(from string: String) -> UIImage? {
-        let data = string.data(using: String.Encoding.ascii)
-        
-        if let filter = CIFilter(name: "CIQRCodeGenerator") {
-            filter.setValue(data, forKey: "inputMessage")
-            let transform = CGAffineTransform(scaleX: 3, y: 3)
-            
-            if let output = filter.outputImage?.transformed(by: transform) {
-                return UIImage(ciImage: output)
-            }
-        }
-        return nil
     }
 }
